แก้เสร็จ 95%

+++ app/Http/Controllers/Admin/ProductController.php
@@ -127,12 +127,24 @@ class ProductController extends Controller
     public function delete($id)
     {
         $product = Product::find($id);
-        unlink(public_path('images/product_image/'.$product->image));
+
+        if ($product->order_item) {
+            session()->flash("error", "ไม่ให้ลบ!");
+
+            return redirect('admin/dashboard');
+        }
+
+
+        $checkFile = is_file(public_path('images/product_image/' . $product->image));
+        // dd($checkFile);
+        if ($checkFile) {
+            unlink(public_path('images/product_image/' . $product->image));
+        }
         // $exists = Storage::disk('local')->exists("public/product_image/" . $product->image); //เจอไฟล์ภาพชื่อเกมือนกัน
         // if ($exists) {
         //     Storage::delete("public/product_image/" . $product->image);
         // }
-        Product::destroy($id);
+        $product->delete();
         session()->flash("success", "ลบข้อมูลเรียบร้อยแล้ว!");
         return redirect('admin/dashboard');
     }
